code_generator/requirements.txt
jinja2
beautifulsoup4
pyyaml
black